Decode HTML entities for pet name
Decode HTML entity and protect from XSS  from pet name since it is a user entered input.
