+++ src/components/pet-display-info/Description.tsx
@@ -1,18 +1,11 @@
 import { Typography, Link as MaterialLink } from "@mui/material";
 
+import {decodeHtmlEntityString} from "../../utils/decodeHtmlEntity.ts";
+
 type Props = {
   description: string | null;
   url: string;
 };
-
-function decodeHtmlEntityString(encodedStr: string | null): string {
-  if (!encodedStr) {
-    return "No description provided.";
-  }
-  const tempElement = document.createElement("div");
-  tempElement.innerHTML = encodedStr;
-  return tempElement.textContent as string;
-}
 
 export default function Description(props: Props) {
   const decodedDescription = decodeHtmlEntityString(props.description);
